requirements.txt
aiogram~=2.11.2
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
colorlog~=5.0.1
cycler==0.11.0
fonttools==4.29.0
frozenlist==1.3.0
idna==3.3
kiwisolver==1.3.2
matplotlib~=3.3.4
multidict==6.0.2
numpy==1.22.1
packaging==21.3
pandas~=1.2.3
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests~=2.25.1
six==1.16.0
urllib3==1.26.8
yarl==1.7.2
tqdm~=4.59.0
scipy~=1.6.1
timebudget~=0.7.1
PyYAML~=5.4.1